feat(cloudflare/nextjs): build: { command } + packager-aware default; document both clouds
Cloudflare parity with the AWS surface:

```ts
const site = yield* Cloudflare.Website.Nextjs("Site", {
  build: { command: "npx next build --no-lint" },
});
```

- build: { command } on Cloudflare.Website.Nextjs (precedence over
  nextjs.buildCommand; open-next.config.ts buildCommand still wins) —
  the inherited Worker rolldown `build` prop is omitted, the composite's
  build IS the OpenNext pipeline
- runner.mjs default is now packager-aware (bunx/npx/yarn/pnpm exec via
  nearest-lockfile detection, sync mirror of the shared packager.ts)
- shared detection extracted to nextjs/packager.ts (aws.ts re-exports)
- 'Configure the build command' sections added to the AWS and Cloudflare
  Next.js frontend docs
