+++ packages/alchemy/src/Cloudflare/Website/Nextjs.ts
@@ -37,9 +37,10 @@ export interface NextjsBuildOptions {
   configPath?: string;
   /**
    * The command the OpenNext pipeline runs to build the Next.js app.
-   * A `buildCommand` set in the project's `open-next.config.ts` takes
-   * precedence over this option.
-   * @default "npx next build"
+   * Prefer the top-level `build: { command }` prop, which takes precedence
+   * over this option; a `buildCommand` set in the project's
+   * `open-next.config.ts` takes precedence over both.
+   * @default `next build` via the detected package runner (bunx/npx/yarn/pnpm exec)
    */
   buildCommand?: string;
   /**
@@ -76,7 +77,14 @@ export interface NextjsProps<
   Bindings extends WorkerBindingProps = {},
 > extends Omit<
   WorkerProps<Bindings>,
-  "vite" | "main" | "assets" | "script" | "bundle" | "source" | "rules"
+  | "vite"
+  | "main"
+  | "assets"
+  | "script"
+  | "build"
+  | "bundle"
+  | "source"
+  | "rules"
 > {
   /**
    * The Next.js project root (the directory containing `next.config.*` and
@@ -91,6 +99,19 @@ export interface NextjsProps<
    * `include`/`exclude` globs when the default is too broad.
    */
   memo?: MemoOptions;
+  /**
+   * Build configuration for the OpenNext build.
+   */
+  build?: {
+    /**
+     * Command that builds the Next.js app, run from `rootDir` (e.g.
+     * `"npx next build --turbopack"`). Takes precedence over
+     * `nextjs.buildCommand`; a `buildCommand` set in the project's
+     * `open-next.config.ts` takes precedence over both.
+     * @default `next build` via the detected package runner (bunx/npx/yarn/pnpm exec)
+     */
+    command?: string;
+  };
   /** Next.js/OpenNext-specific build and dev configuration. */
   nextjs?: NextjsBuildOptions;
   /**
@@ -251,14 +272,15 @@ export interface NextjsProps<
  * ```
  *
  * @section Build Configuration
- * The `nextjs` prop tunes the OpenNext pipeline: a custom build command,
- * minification, or reusing an existing `.next` build.
+ * `build: { command }` sets the build command; the `nextjs` prop tunes the
+ * rest of the OpenNext pipeline (minification, reusing an existing `.next`
+ * build).
  *
  * @example Minified build with a custom command
  * ```typescript
  * const site = yield* Cloudflare.Website.Nextjs("Site", {
+ *   build: { command: "npx next build --no-lint" },
  *   nextjs: {
- *     buildCommand: "npx next build --no-lint",
  *     minify: true,
  *   },
  * });
@@ -318,7 +340,10 @@ export const Nextjs: {
         Effect.map(
           Effect.isEffect(propsEff) ? propsEff : Effect.succeed(propsEff),
           (props) => ({
-            ...props,
+            // `build` is the OpenNext build config (forwarded through the
+            // source provider below), not the Worker's rolldown options —
+            // keep it out of the Worker props.
+            ...(({ build: _build, ...worker }) => worker)(props ?? {}),
             // OpenNext's revalidation queues (memory-queue, do-queue) fetch
             // the worker back through `WORKER_SELF_REFERENCE`. Always wire
             // the self service binding — it's inert when unused, and its
@@ -360,6 +385,9 @@ export const Nextjs: {
                         ? { dev: { mode: devMode } }
                         : {}),
                     }))(props.nextjs)
+                  : {}),
+                ...(props?.build?.command !== undefined
+                  ? { buildCommand: props.build.command }
                   : {}),
               },
             },
