+++ packages/frontend-frameworks/src/nextjs/Nextjs.ts
@@ -62,7 +62,7 @@ export interface NextjsFrameworkOptions {
          * The command the OpenNext pipeline runs to build the Next.js app.
          * A `buildCommand` set in the project's `open-next.config.ts` takes
          * precedence over this option.
-         * @default "npx next build"
+         * @default `next build` via the detected package runner (bunx/npx/yarn/pnpm exec)
          */
         readonly buildCommand?: string | undefined;
         /** Skip the internal `next build` (reuse an existing `.next`). */
